+++ src/nrmc/precinct_to_precinct.py
@@ -6,7 +6,8 @@ import random
 
 
 from .core import MetropolisProcess, TemperedProposalMixin, compute_dot_product, exp
-from .state import update_contested_edges, simply_connected, check_population
+from src.nrmc.constraints import simply_connected
+from src.nrmc.updaters import update_contested_edges, check_population
 
 
 class PrecintFlow(MetropolisProcess):
@@ -86,81 +87,5 @@ class PrecintFlowTempered(PrecintFlow, TemperedProposalMixin):
     pass # woohoo! no code to write here
 
 
-    # def proposal(self, state):
-    #
-    #     # get all the proposals, with their associated probabilities
-    #     proposals = self.get_proposals(state, filter_connected=False)
-    #     # pick a proposal
-    #     try:
-    #         proposal, score = self.pick_proposal(proposals)
-    #     except ValueError:
-    #         # state.involution *=-1
-    #         # self._proposal_state.involution *= -1
-    #         # return self.proposal(state) # if we can't find a valid proposal
-    #         return (None, None, None), 0 # we couldn't find a valid proposal, this is guaranteed to cause involution
-    #
-    #     node_id, old_color, new_color = proposal
-    #     self._proposal_state.handle_move(proposal)
-    #     reverse_proposals = self.get_proposals(self._proposal_state, filter_connected=True)
-    #     reverse_proposal_probs = {k:self.score_to_proposal_prob(v) for k,v in reverse_proposals.items()}
-    #     proposal_probs = {k: self.score_to_proposal_prob(v) for k,v in proposals.items()}
-    #
-    #     try:
-    #         q_prime = reverse_proposals[(node_id, new_color, old_color)]/sum(reverse_proposal_probs.values())
-    #         q = proposals[proposal]/sum(proposal_probs.values())
-    #         return proposal, q_prime/q*exp(-score*self.beta*self.measure_beta) # TODO should this be a different func?
-    #     except KeyError:
-    #         return proposal, 0 # sometimes reverse is not contained in proposals list
 
-
-    # def handle_rejection(self, prop, state):
-    #     super().handle_rejection(prop, state)
-    #     if prop[0] is not None: # we couldn't find a valid proposal
-    #         self._proposal_state.handle_move((prop[0], prop[2], prop[1])) # flip back to old color
-    #     self._proposal_state.involution *= -1
-
-    # def pick_proposal(self, proposals):
-    #
-    #     # threshold = np.random.random()
-    #     # keys = sorted(proposals.keys())
-    #     counter = 0
-    #     while True:
-    #         proposal = random.choices(list(proposals.keys()), weights=self.score_to_proposal_prob(proposals.values()), k=1)[0]
-    #         if self.check_connected(self.state, proposal[0], proposal[1], proposal[2]):
-    #             return proposal, proposals[proposal]
-    #         else:
-    #             counter +=1
-    #             proposals[proposal] = 0 # this wasn't connected, so we can't pick it
-    #             if counter >= len(proposals):
-    #                 # return (None, None, None), 0 # not actually a valid proposal, guaranteed to result in involution
-    #                 raise ValueError("Could not find connected proposal")
-    #
-    #             proposals[proposal] = 0 # probability is zero if it's not connected
-
-    # def get_proposals(self, state=None, filter_connected=False):
-    #
-    #     update_contested_edges(state)
-    #     # update_perimeter_aggressive(state)
-    #     proposals = defaultdict(int)
-    #
-    #     directed_edges = self.get_directed_edges(state)
-    #     for iedge in directed_edges:
-    #
-    #         old_color = state.node_to_color[iedge[0]]  # find the district that the original belongs to
-    #         new_color = state.node_to_color[iedge[1]]
-    #
-    #         if not check_population(state, old_color, minimum=500):
-    #             continue # ensure that population of either does not drop below 500
-    #             # TODO parameterize this
-    #
-    #         if (iedge[0], old_color, new_color) in proposals:
-    #             continue # we've already scored this proposal, no need to redo
-    #
-    #         if filter_connected and not self.check_connected(state, iedge[0], old_color, new_color):
-    #              continue
-    #
-    #         score = self.score_proposal(iedge[0], old_color, new_color, state)  # smaller score is better
-    #         proposals[(iedge[0], old_color, new_color)] = score
-    #
-    #     return proposals
