+++ src/nrmc/single_node_flip.py
@@ -2,7 +2,8 @@ import random
 import copy
 
 from .core import MetropolisProcess, TemperedProposalMixin
-from .state import update_contested_edges, check_population, simply_connected
+from src.nrmc.constraints import simply_connected
+from src.nrmc.updaters import update_contested_edges, check_population
 from .scores import cut_length_score
